requirements.txt
asgiref==3.6.0
Django==4.2
django-factory-boy==1.0.0
factory-boy==3.2.1
Faker==18.4.0
python-dateutil==2.8.2
six==1.16.0
sqlparse==0.4.4
